% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_cjpg_tidy}
\alias{da_cjpg_tidy}
\title{da_cjpg_tidy}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 338 rows and 7 columns.
}
\usage{
da_cjpg_tidy
}
\description{
da_cjpg_tidy
}
\keyword{datasets}
